requirements.txt
torch==1.12.1
pandas==2.0.3
numpy==1.24.3
scipy==1.10.1
matplotlib==3.7.2
seaborn==0.12.2
hydra-core==1.3.2
wandb==0.16.1
tqdm==4.65.0